backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
pillow==9.5.0
requests==2.30.0
pyngrok==6.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
transformers==4.35.2
accelerate==0.25.0
bitsandbytes==0.41.1
huggingface-hub